+++ src/FrameThread.h
@@ -27,7 +27,7 @@ public:
   void setPreviewChannel(std::string channelName);
 
 protected:
-  bool writeAll(uint32_t file, const uint8_t* buffer, uint32_t length);
+  bool writeAll(uint64_t file, const uint8_t* buffer, uint32_t length);
 
 private:
   std::shared_ptr<FfmpegProcess> ffmpegProcess;
